:recycle: refactor: auth method
